Fix channels 128 cable (#23)
* fix to cable 128

* fix to cable 128

* sa

* update bs

---------

+++ src/main/scala/net/bdew/ae2stuff/items/visualiser/data.scala
@@ -26,7 +26,7 @@ case class VNode(x: Int, y: Int, z: Int, flags: VNodeFlags.ValueSet)
 case class VLink(
     node1: VNode,
     node2: VNode,
-    channels: Byte,
+    channels: Int,
     flags: VLinkFlags.ValueSet
 )
 
@@ -54,7 +54,7 @@ class VisualisationData(var nodes: Seq[VNode], var links: Seq[VLink])
       links = for (x <- 0 until linkCount) yield {
         val n1 = in.readInt()
         val n2 = in.readInt()
-        val c = in.readByte()
+        val c = in.readInt()
         val f = in.readByte()
         VLink(
           nodes(n1),
@@ -82,7 +82,7 @@ class VisualisationData(var nodes: Seq[VNode], var links: Seq[VLink])
     for (l <- links) {
       out.writeInt(nodeMap(l.node1))
       out.writeInt(nodeMap(l.node2))
-      out.writeByte(l.channels)
+      out.writeInt(l.channels)
       out.writeByte(l.flags.toBitMask(0).toByte)
     }
   }
